Add system parameters for commitment
The parameters needed for commitments to blinded attributes are now
included in the system parameters. Because we are no longer compatible
with libsignal, remove the relevant functions and test case.

Another change is that the system parameters are now generated from only
one customization label. This label should be used as a
nothing-up-my-sleeve number.